Create the 3.0 release. (#43)
* Create the 3.x release series.

We're starting to have submissions for new collections for ansible-3.0.
Need to make the ./3/ directory and ansible.in so that those have a place
to land.

* Use semantic versioning

The next version of ansible moves to a semantic version model.  So the
major release will be 3.0.0 and the minor releases that come out every
three weeks will be 3.1.0, 3.2.0, etc.