CI: Use version 5.10.0 to test builds in CI
The purpose of the job is to build what would be the next version of
ansible, including the changes of the PR, to see if it works.

We were trying to build ansible 5.0.0 in the job because that's what the
role defaults to.

Specify a future (unreleased) version so we can test the build properly.